Fix frontend Symbols & Candles page crash, normalize fields, and add regression test

+++ backend/app/schemas/market_data.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -19,10 +20,10 @@ class SymbolResponse(BaseModel):
     base_asset: str
     quote_asset: str
     trading_status: str
-    tick_size: float
-    step_size: float
-    minimum_quantity: float
-    minimum_notional: float
+    tick_size: Decimal
+    step_size: Decimal
+    minimum_quantity: Decimal
+    minimum_notional: Decimal
     price_precision: int
     quantity_precision: int
     refreshed_at: datetime
@@ -33,21 +34,21 @@ class CandleResponse(BaseModel):
     timeframe: str
     open_time: datetime
     close_time: datetime
-    open_price: float
-    high_price: float
-    low_price: float
-    close_price: float
-    volume: float
-    quote_volume: float
+    open_price: Decimal
+    high_price: Decimal
+    low_price: Decimal
+    close_price: Decimal
+    volume: Decimal
+    quote_volume: Decimal
     trade_count: int
 
 
 class SnapshotResponse(BaseModel):
     symbol: str
-    last_price: float
-    bid_price: float
-    ask_price: float
-    bid_quantity: float
-    ask_quantity: float
-    spread_bps: float
+    last_price: Decimal
+    bid_price: Decimal
+    ask_price: Decimal
+    bid_quantity: Decimal
+    ask_quantity: Decimal
+    spread_bps: Decimal
     snapshot_at: datetime
